+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Switch OSD to Dear ImGui
 * Default monotonic time to 1s to prevent weird behaviours on some Unity games
 * Improve the way to detect the location of SDL_DYNAPI_entry address
+* Improve "Skip unmapped pages" by detecting uncommitted segments
 
 ### Fixed